[llava] Use huggingface LLaVA instead of depending on third-party/LLaVa
Currently we depend on third-party/LLaVA for llava model definition.
This is hard to use because we have to pull LLaVA in as a git submodule
and install from there. It also breaks a lot of dependency assumptions.

This PR removes `third-party/LLaVA`, in favor of huggingface llava model
definition.

ghstack-source-id: 4714a934127f3909ae73c9112358bbe1818fa0b7
Pull Request resolved: https://github.com/pytorch/executorch/pull/4687